Create Day-1
You might be aware that the events of the Star Wars saga took place a long time ago, in a galaxy that was rather far away. But just how far away was it? 
We can try to describe it by repeating the word "far" a certain number of times in the following sentence format:
A long time ago in a galaxy far, far away...

In the above example, the word far is repeated twice. However, we'd instead like to repeat it exactly 1<= N <= 5 times without changing the rest of the sentence at all. There should be a comma right after each occurrence except for the last one. 

Given N, can you produce the correct sentence? 

from https://dmoj.ca/problem/wc15c2j1