Determine the size of the smallest neighborhood
Language: Python 3.11

First the number of villages "numVillages" is read from the input. Using the first for loop, each position is read and stored in a list named "position".

To calculate the size of the neighbors, the position list is sorted with "position.sort()".

The next for loop calculates the neighborhood size starting from the second element of the list and ending with the n-1 - leftmost and rightmost villages have only one neighbor.

The sizes are stored in a list called "neighborhood".

The smallest neighborhood is returned by the built-in "min" function.

Source: https://dmoj.ca/problem/ccc18s1