Find out if students have raised enough money
Language: Python 3.11

Read all inputs.
The student's proportion 'proportion' is a string and will be split.

Then iterate over all elements in the 'propotion' list and convert them to floats. Now they can be multiplied by the number of students in the brunch 'students' and stored in the list 'stud_per_year'. We use 'int()' to round the number of students per year.

By adding up all the elements of 'stud_per_year' we can find out if there are any students missing. Get the index of the group with the highest percentage of students and add the difference.

Now calculate the total money raised at the brunch by multiplying each element of 'stud_per_year' by the money they have to pay.

We compare if half of the money raised is more than what the trip cost 'cost' and print the answer

All of this is repeated 10 times for the ten cases.

Source: https://dmoj.ca/problem/ecoo17r1p1