Order the songs in the current playlist
Language: Python 3.11

Define a string with the names of the songs 'songs' and start the varibale 'button' with a number other than 4.

As long as 'button' reads a value different than 4, the while loop runs.

Compare the value of button to decide how to change the order of the characters in 'songs'.

Create a new empty string called 'output', where each pair of songs is enclosed as a character using a for loop and separated by a space.

Source: https://dmoj.ca/problem/ccc08j2