Determine the total number of bonuses awarded
Language: Python 3.11

First we read the inputs and store them as two different integers 'f' and 'd'.

The first for loop stores the sales per day for all franchisees in the list 'sales'. These need to be converted to an integer. An iteration and coverage of each element is done by the for loop directly after the values of that day are appended to 'sales'.

By adding all the values of the rows of 'sales', we obtain the total sales per day 'sum_per_day'. If the remainder of the division between this sum and 13 is zero, it's a multiple of 13. The if statement is true, so we add the result to the variable 'bonus'. Use '//' to get an integer.

The same process happens for the columns of 'sales'. Here we get the sum of the sales per franchisee 'sum_per_franchisee '.

Then print the final value of 'bonus' before moving on to the next test-case (inside the first for-loop).

Source: https://dmoj.ca/problem/ecoo17r3p1