Determine Wesley's final score
Language: Python 3.11

Assuming that the inputs meet the requirements, the inputs are read and stored as variables and/or lists.

The number of points 'total_points' starts with a value equal to the length of the sequence 'S', because for each move Wesley becomes a point. The points from the combos will be added later.

We have to compare all moves of the sequence with the combos and see if they match. For this we use a while loop. The loop runs as long as we haven't compared all moves, so we use the length of the sequence 'len(S)' to guarantee it.

We compare each group of moves with each combo, starting with position 0. They have to be the same length. It's also important that the sequence of moves matches the longest combo. That's why we use an if statement and additionally compare the current combo length 'combo_len' with the longest of all: 'longest_match_lengths'. If it's true, the new value of 'longest_match_length' is equal to 'combo_len' and we store the point of this combo in the variable 'best_points'.

After comparing all of them, , we add 'best_points' to 'total_points' and skip the next moves that coincide with the combo and start comparing again. If not, we continue with the next one, adding 1 to the counter 'j'.

Source: https://dmoj.ca/problem/wac3p3