Get the num of occupied parking spaces on both days
Language: Python 3.11

First the program reads the inputs and stores them as 'parkingSpaces', 'yesterday' and 'today' according to the problem specification.

To calculate the number of parking spaces occupied yesterday and today, we use a counter 'occupied' and start it equal to 0.

Then we compare each character of yesterday's and today's information using a for loop. If both are equal to C, so the parking space is occupied, we add 1 to 'occupied'.

After comparing them all, we print the final value of 'occupied'.

Source: https://dmoj.ca/problem/ccc18j2