Tell Cezar when to draw a card.
Language: Python 3.11

Read the number of cards Cezar has drawn 'N' and write the value of each card in the 'drawn' list.

Then find the difference 'x' from the sum of the cards so far to 21.

Calculate the number of remaining cards that are less than or equal to 'x'. This depends on which cards he has already drawn. For example, if 'x' is less than or equal to 9, the number of remaining cards in the deck is equal to the number of cards from 1 to 'x' minus the number of cards with those values that he's already drawn. Therefore, we need a 'count' counter to determine how many of the cards he has drawn fall into this category. The same goes for the other cases. If 'x' is equal to 10, it means that we have to take into account all the numbered cards (9 cards in each of the 4 suits) and the Kings, Queens and Jacks, one per suit. If it's more than 10, it means that this time even the Aces are included.

If the difference between the 52 cards, the number of cards remaining in the deck and the number of cards drawn is equal to or greater than the number of cards remaining, print DOSTA. In the other case, print VUCI.

PS: This code was correct in 10 out of 11 tests. It must be checked again for Batch 10.

Source: https://dmoj.ca/problem/coci17c1p1