Output the card game score
Language: Python 3.11

The first step is to read the types of cards in the deck and store them in the list 'cards'.

To determine if none of the next cards in the deck after a high card is a high card, we define the function 'no_high'. If the value matches one of the high cards, the function returns False. Otherwise it returns True.

We need to compare each of the cards a player takes to see if a player scores. To do this we use a for loop. To score, the player must meet the following conditions: He must take a high card, the remaining cards must be sufficient according to the rules, and they must not be high cards.

Player A starts the game, then B, then A and so on. This means that for the number of turns 'j' starting from zero that are multiples of two, it's the turn of player A. When a player scores, the value of 'points' changes and we print how many points the player scored that time and add this number to the total score 'sum_A' or 'sum_B'.

At the end we print the total score.

Source: https://dmoj.ca/problem/ccc99s1