Find out if Lena can organize the boxes
Language: Python 3.11

We use functions to solve this problem. These are defined at the top of the code. The main program calls each of them when needed to solve the problem.

Each function has its own description for better understanding.

Because the problem description says so, we don't consider the case when there is an empty box.

This problem is from Timus Judge - Problem 2144

Source: https://acm.timus.ru/problem.aspx?space=1&num=2144