Distinguish English text from French text
Language: Python 3.11

First we read the number of lines 'numlines'. We store the input lines in the string 'text' and convert it to lowercase.

Now that we have the hole text in lowercase, we compare if the number of times the letter t is used is more than the number of times the letter s is used. For this we use the 'count()' function. If that's the case, we can assume that the text is written in English and print this answer. If that's not the case, we print French.

Source: https://dmoj.ca/problem/ccc11s1