Calculate the streams at the foot of the mountain
Language: Python 3.11

We use the input, the initial number of streams 'initialState', to determine how many lines will come next. The first states are stored in the list 'streams'.

Now we need to determine which streams have been split or merged until we read the number 77. That's why we use a while loop. As long as the condition is false, the program runs. To find out, we use 'indication' to identify the different possibilities. If 'indication' is equal to 77, the variable 'done' is true and the program is done.

If it's 99, the flow splits. This means that we have to calculate the new values of the two new streams. We take the flow at the given 'position' and calculate the value by multiplying it by the 'percentage'. We use the 'round()' function to get an integer. The new stream on the left is equal to the difference between the value before the splitting and the new value of the stream on the right. We add it to 'streams' with 'insert()' function, specifying its location.

If 'indication' is 88, we only need to add the stream at the index 'position - 1' and 'position'. This new value will replace the right stream (index position-1). We don't need the left stream because it doesn't exist anymore, so we have to delete it by using the function 'pop()' and indicating its index.

At the end of the program we print 'streams'. The list contains the current state of the streams. By using * we can print the values next to each other, e.g. 5 55, otherwise the output will be in the form of a list, e.g. [5, 55].

Source: https://dmoj.ca/problem/ccc00s2